requirements.txt
python-dotenv
azure-identity
azure-ai-projects==2.0.0b1
openai